requirements.txt
aiofiles==23.2.1
aiogram==3.0.0b5
aiohttp==3.8.3
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
blinker==1.6.3
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
curl-cffi==0.5.9
duckduckgo-search==3.9.3
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.3.3
g4f==0.1.7.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
magic-filter==1.0.9
MarkupSafe==2.1.3
multidict==6.0.4
Naked==0.1.32
nest-asyncio==1.5.8
numpy==1.24.4
openai==0.26.4
Pillow==10.1.0
platformdirs==3.11.0
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==1.10.4
PyExecJS==1.5.1
pyjsparser==2.7.1
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.2
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tiktoken==0.5.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.14
waitress==2.1.2
websockets==12.0
Werkzeug==3.0.0
win32-setctime==1.1.0
yarl==1.8.2
zipp==3.17.0
